+++ docker/chain-indexer.conf
@@ -6,7 +6,11 @@ uexplorer {
         // polling new blocks after sync is done from local node otherwise from the entire peer network (better resiliency)
         peer-address-to-poll-from = "http://213.239.193.208:9053"
 
-        backend-type = { type = "cassandra-db", parallelism = ${BACKEND_INDEXING_PARALLELISM} }
+        backend-type = { type = "no-backend" }
+        // backend-type = { type = "cassandra-db", parallelism = ${BACKEND_INDEXING_PARALLELISM} }
+
+        graph-backend-type = { type = "no-graph-backend" }
+        // graph-backend-type = { type = "janus-graph" }
     }
 }
 datastax-java-driver.basic.contact-points = [${CASSANDRA_SEED_HOST}":"${CASSANDRA_SEED_PORT}]
