+++ docker/drain-cassandra.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-echo "Flushing Cassandra memtables to disk and then stopping and removing all containers"
+echo "Flushing Cassandra memtables to disk ..."
 
 docker exec -it cassandra nodetool drain
 docker exec -it stargate nodetool drain
