+++ modules/alert-plugin/src/main/scala/org/ergoplatform/uexplorer/plugin/alert/Analyzer.scala
@@ -1,22 +1,21 @@
 package org.ergoplatform.uexplorer.plugin.alert
 
 import org.apache.tinkerpop.gremlin.process.traversal.dsl.graph.GraphTraversalSource
-import org.ergoplatform.uexplorer.SortedTopAddressMap
+import org.ergoplatform.uexplorer.{SortedTopAddressMap, Storage}
 import org.ergoplatform.uexplorer.plugin.alert.HighValueDetector.{BlockMatch, TxMatch}
-import org.ergoplatform.uexplorer.utxo.UtxoState
 
 trait Analyzer {
 
   def trackTx(
-    txMatch: TxMatch,
-    utxoState: UtxoState,
-    graphTraversalSource: Option[GraphTraversalSource]
+               txMatch: TxMatch,
+               utxoState: Storage,
+               graphTraversalSource: Option[GraphTraversalSource]
   ): Option[TxMatch]
 
   def trackBlock(
-    blockMatch: BlockMatch,
-    utxoState: UtxoState,
-    graphTraversalSource: Option[GraphTraversalSource]
+                  blockMatch: BlockMatch,
+                  utxoState: Storage,
+                  graphTraversalSource: Option[GraphTraversalSource]
   ): Option[BlockMatch]
 
 }
