+++ modules/alert-plugin/src/main/scala/org/ergoplatform/uexplorer/plugin/alert/Detector.scala
@@ -5,21 +5,20 @@ import org.ergoplatform.uexplorer.db.BestBlockInserted
 import org.ergoplatform.uexplorer.node.ApiTransaction
 import org.ergoplatform.uexplorer.plugin.alert.Detector.AlertMessage
 import org.ergoplatform.uexplorer.plugin.alert.HighValueDetector.{BlockMatch, TxMatch}
-import org.ergoplatform.uexplorer.utxo.UtxoState
-import org.ergoplatform.uexplorer.{Address, BoxId, SortedTopAddressMap, TopAddressMap}
+import org.ergoplatform.uexplorer.{Address, BoxId, SortedTopAddressMap, Storage, TopAddressMap}
 
 trait Detector {
 
   def inspectNewPoolTx(
-    tx: ApiTransaction,
-    utxoState: UtxoState,
-    graphTraversalSource: Option[GraphTraversalSource]
+                        tx: ApiTransaction,
+                        utxoState: Storage,
+                        graphTraversalSource: Option[GraphTraversalSource]
   ): List[TxMatch]
 
   def inspectNewBlock(
-    newBlock: BestBlockInserted,
-    utxoState: UtxoState,
-    graphTraversalSource: Option[GraphTraversalSource]
+                       newBlock: BestBlockInserted,
+                       utxoState: Storage,
+                       graphTraversalSource: Option[GraphTraversalSource]
   ): List[BlockMatch]
 }
