+++ modules/alert-plugin/src/main/scala/org/ergoplatform/uexplorer/plugin/alert/HighValueDetector.scala
@@ -1,12 +1,11 @@
 package org.ergoplatform.uexplorer.plugin.alert
 
-import org.ergoplatform.uexplorer.db.{BestBlockInserted, Block}
 import org.ergoplatform.uexplorer.*
 import org.ergoplatform.uexplorer.node.ApiTransaction
 import org.ergoplatform.uexplorer.plugin.alert.Detector.AlertMessage
 import HighValueDetector.*
 import org.apache.tinkerpop.gremlin.process.traversal.dsl.graph.GraphTraversalSource
-import org.ergoplatform.uexplorer.utxo.UtxoState
+import org.ergoplatform.uexplorer.db.{BestBlockInserted, Block}
 import org.slf4j.{Logger, LoggerFactory}
 
 import java.text.DecimalFormat
@@ -18,26 +17,26 @@ class HighValueDetector(txErgValueThreshold: Long, blockErgValueThreshold: Long)
 
   def inspectNewPoolTx(
     tx: ApiTransaction,
-    utxoState: UtxoState,
+    storage: Storage,
     graphTraversalSource: Option[GraphTraversalSource]
   ): List[TxMatch] = {
     def sumAddressValues(addresses: Set[Address]): Map[Address, Address.State] =
       addresses
-        .flatMap(a => utxoState.getUtxosByAddress(a).map(a -> _))
+        .flatMap(a => storage.getUtxosByAddress(a).map(a -> _))
         .foldLeft(Map.empty[Address, Address.State]) { case (acc, (address, valueByBox)) =>
-          acc.updated(address, Address.State(valueByBox.values.sum, utxoState.getAddressStats(address)))
+          acc.updated(address, Address.State(valueByBox.values.sum))
         }
 
     val outputsWithoutPaybacksAndFees =
       tx.outputs.filterNot(o =>
         tx.inputs.exists(i =>
-          utxoState.getAddressByUtxo(i.boxId).contains(o.address)
+          storage.getAddressByUtxo(i.boxId).contains(o.address)
         ) || o.address == Const.FeeContract.address
       )
     Option(outputsWithoutPaybacksAndFees.iterator.map(_.value).sum)
       .filter(_ >= txErgValueThreshold * Const.NanoOrder)
       .map { value =>
-        val inputAddresses       = tx.inputs.iterator.map(_.boxId).flatMap(utxoState.getAddressByUtxo).toSet
+        val inputAddresses       = tx.inputs.iterator.map(_.boxId).flatMap(storage.getAddressByUtxo).toSet
         val inputValueByAddress  = sumAddressValues(inputAddresses)
         val outputAddresses      = outputsWithoutPaybacksAndFees.iterator.map(_.address).toSet
         val outputValueByAddress = sumAddressValues(outputAddresses)
@@ -60,7 +59,7 @@ class HighValueDetector(txErgValueThreshold: Long, blockErgValueThreshold: Long)
 
   def inspectNewBlock(
     newBlock: BestBlockInserted,
-    utxoState: UtxoState,
+    storage: Storage,
     graphTraversalSource: Option[GraphTraversalSource]
   ): List[BlockMatch] = List.empty
 
@@ -89,10 +88,7 @@ object HighValueDetector {
       val fmtOutputAddrSum = valueFormat.format(outputs.valuesIterator.map(_._1).sum / Const.NanoOrder)
 
       def stringify(tuple: (Address, Address.State)) = tuple match {
-        case (address, Address.State(value, Some(Address.Stats(lastTxHeight, txCount, boxCount)))) =>
-          val fmtVal = valueFormat.format(value / Const.NanoOrder)
-          s"${address.asInstanceOf[String].take(5)} $fmtVal @ [$lastTxHeight/$txCount/$boxCount]"
-        case (address, Address.State(value, None)) =>
+        case (address, Address.State(value)) =>
           val fmtVal = valueFormat.format(value / Const.NanoOrder)
           s"${address.asInstanceOf[String].take(5)} $fmtVal"
       }
