+++ modules/alert-plugin/src/main/scala/org/ergoplatform/uexplorer/plugin/alert/SourceAnalyzer.scala
@@ -2,9 +2,8 @@ package org.ergoplatform.uexplorer.plugin.alert
 
 import org.apache.tinkerpop.gremlin.process.traversal.dsl.graph.GraphTraversalSource
 import org.apache.tinkerpop.gremlin.structure.io.graphml.GraphMLWriter
-import org.ergoplatform.uexplorer.SortedTopAddressMap
+import org.ergoplatform.uexplorer.{SortedTopAddressMap, Storage}
 import org.ergoplatform.uexplorer.plugin.alert.HighValueDetector.{BlockMatch, TxMatch}
-import org.ergoplatform.uexplorer.utxo.UtxoState
 
 import java.io.FileOutputStream
 
@@ -12,7 +11,7 @@ class SourceAnalyzer extends Analyzer {
 
   def trackTx(
     txMatch: TxMatch,
-    utxoState: UtxoState,
+    storage: Storage,
     graphTraversalSource: Option[GraphTraversalSource]
   ): Option[TxMatch] =
     // TODO currently lack of options for visualizing graphson/graphml besides https://gitlab.com/ouestware/retina
@@ -20,7 +19,7 @@ class SourceAnalyzer extends Analyzer {
 
   def trackBlock(
     blockMatch: BlockMatch,
-    utxoState: UtxoState,
+    storage: Storage,
     graphTraversalSource: Option[GraphTraversalSource]
   ): Option[BlockMatch] = Option.empty
