+++ modules/cassandra/src/main/scala/org/ergoplatform/uexplorer/cassandra/entity/CassandraHeaderWriter.scala
@@ -7,11 +7,10 @@ import com.datastax.oss.driver.api.core.cql.{BoundStatement, PreparedStatement}
 import com.datastax.oss.driver.api.core.data.UdtValue
 import com.datastax.oss.driver.internal.core.`type`.UserDefinedTypeBuilder
 import com.typesafe.scalalogging.LazyLogging
-import org.ergoplatform.uexplorer.db.Block
 import org.ergoplatform.uexplorer.cassandra.CassandraBackend
 import eu.timepit.refined.auto.*
 import org.ergoplatform.uexplorer.cassandra
-import org.ergoplatform.uexplorer.db.BestBlockInserted
+import org.ergoplatform.uexplorer.db.{BestBlockInserted, Block}
 
 trait CassandraHeaderWriter extends LazyLogging { this: CassandraBackend =>
   import Headers._
