+++ modules/cassandra/src/main/scala/org/ergoplatform/uexplorer/cassandra/entity/CassandraTokensWriter.scala
@@ -4,12 +4,11 @@ import akka.NotUsed
 import akka.stream.scaladsl.Flow
 import com.datastax.oss.driver.api.core.cql.{BoundStatement, DefaultBatchType, PreparedStatement}
 import com.typesafe.scalalogging.LazyLogging
-import org.ergoplatform.uexplorer.db.Block
 import org.ergoplatform.uexplorer.cassandra.CassandraBackend
 import eu.timepit.refined.auto.*
+import org.ergoplatform.uexplorer.db.{BestBlockInserted, Block}
 
 import scala.collection.immutable.ArraySeq
-import org.ergoplatform.uexplorer.db.BestBlockInserted
 
 trait CassandraTokensWriter extends LazyLogging {
   this: CassandraBackend =>
