+++ modules/chain-indexer/src/main/resources/application.conf
@@ -52,12 +52,12 @@ uexplorer {
         // polling new blocks after sync is done from local node otherwise from the entire peer network (better resiliency)
         peer-address-to-poll-from = "http://213.239.193.208:9053"
 
-        backend-type = { type = "NoBackend"}
-        // backend-type = { type = "cassandra-db", parallelism = 1}
+        backend-type = { type = "no-backend" }
+        // backend-type = { type = "cassandra-db", parallelism = 1 }
         // backend-type = { type = "in-memory-db" }
 
-        graph-backend-type = { type = "NoGraphBackend"}
-        // graph-backend-type = { type = "janus-graph"}
+        graph-backend-type = { type = "no-graph-backend" }
+        // graph-backend-type = { type = "janus-graph" }
         // graph-backend-type = { type = "in-memory-graph" }
 
         protocol {
