+++ modules/chain-indexer/src/test/scala/org/ergoplatform/uexplorer/indexer/SchedulerSpec.scala
@@ -8,7 +8,6 @@ import org.ergoplatform.uexplorer.indexer.chain.*
 import org.ergoplatform.uexplorer.indexer.mempool.{MempoolStateHolder, MempoolSyncer}
 import org.ergoplatform.uexplorer.indexer.mempool.MempoolStateHolder.MempoolState
 import org.ergoplatform.uexplorer.indexer.plugin.PluginManager
-import org.ergoplatform.uexplorer.utxo.{MvUtxoState, UtxoState}
 import org.scalatest.BeforeAndAfterAll
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.freespec.AsyncFreeSpec
@@ -20,13 +19,14 @@ import org.ergoplatform.uexplorer.janusgraph.api.InMemoryGraphBackend
 
 import scala.collection.immutable.{ListMap, TreeMap}
 import scala.concurrent.Future
-import org.ergoplatform.uexplorer.ProtocolSettings
+import org.ergoplatform.uexplorer.{ProtocolSettings, Storage}
 import org.ergoplatform.uexplorer.cassandra.api.InMemoryBackend
 import org.ergoplatform.uexplorer.http.{LocalNodeUriMagnet, Rest, TestSupport}
 import org.ergoplatform.uexplorer.http.RemoteNodeUriMagnet
 import org.ergoplatform.uexplorer.http.BlockHttpClient
 import org.ergoplatform.uexplorer.http.MetadataHttpClient
 import org.ergoplatform.uexplorer.indexer.chain.Initializer.ChainEmpty
+import org.ergoplatform.uexplorer.mvstore.MvStorage
 
 import java.nio.file.Paths
 
@@ -68,25 +68,26 @@ class SchedulerSpec extends AsyncFreeSpec with TestSupport with Matchers with Be
         Response.ok(Rest.blocks.byId(blockId))
     }
 
-  val utxoState     = MvUtxoState().get
+  val storage       = MvStorage().get
   val blockClient   = new BlockHttpClient(new MetadataHttpClient[WebSockets](minNodeHeight = Rest.info.minNodeHeight))
-  val backend       = new InMemoryBackend
-  val graphBackend  = new InMemoryGraphBackend
+  val backend       = Some(new InMemoryBackend)
+  val graphBackend  = Some(new InMemoryGraphBackend)
   val pluginManager = new PluginManager(List.empty)
-  val chainIndexer  = new BlockIndexer(backend, graphBackend, blockClient, utxoState)
+  val blockIndexer  = new BlockIndexer(storage)
+  val chainIndexer  = new ChainIndexer(backend, graphBackend, blockClient, blockIndexer)
   val mempoolSyncer = new MempoolSyncer(blockClient)
-  val initializer   = new Initializer(utxoState, backend, graphBackend)
+  val initializer   = new Initializer(storage, backend, graphBackend)
   val scheduler     = new Scheduler(pluginManager, chainIndexer, mempoolSyncer, initializer)
 
   "Scheduler should sync from 1 to 4150 and then from 4150 to 4200" in {
     initializer.init shouldBe ChainEmpty
     scheduler.periodicSync.flatMap { mempoolState =>
-      utxoState.getLastHeight.get shouldBe 4150
-      utxoState.findMissingHeights shouldBe empty
+      storage.getLastHeight.get shouldBe 4150
+      storage.findMissingHeights shouldBe empty
       mempoolState.stateTransitionByTx.size shouldBe 9
       scheduler.periodicSync.map { newMempoolState =>
-        utxoState.getLastHeight.get shouldBe 4200
-        utxoState.findMissingHeights shouldBe empty
+        storage.getLastHeight.get shouldBe 4200
+        storage.findMissingHeights shouldBe empty
         newMempoolState.stateTransitionByTx.size shouldBe 0
       }
     }
