+++ modules/explorer-core/src/main/scala/org/ergoplatform/uexplorer/Storage.scala
@@ -1,25 +1,29 @@
-package org.ergoplatform.uexplorer.utxo
+package org.ergoplatform.uexplorer
 
 import org.ergoplatform.uexplorer.{Address, BlockId, BlockMetadata, BoxId, Height, Value}
 
 import scala.collection.immutable.TreeSet
 
-trait UtxoState {
+trait Storage {
 
   def isEmpty: Boolean
+
+  def containsBlock(blockId: BlockId, atHeight: Height): Boolean
+
+  def findMissingHeights: TreeSet[Height]
+
+  def getCurrentVersion: Long
+
+  def getBlockById(blockId: BlockId): Option[BlockMetadata]
 
   def getLastHeight: Option[Height]
 
   def getLastBlocks: Map[BlockId, BlockMetadata]
 
-  def getAddressStats(address: Address): Option[Address.Stats]
-
-  def containsBlock(blockId: BlockId, atHeight: Height): Boolean
+  def getBlocksByHeight(atHeight: Height): Map[BlockId, BlockMetadata]
 
   def getAddressByUtxo(boxId: BoxId): Option[Address]
 
   def getUtxosByAddress(address: Address): Option[Map[BoxId, Value]]
-
-  def findMissingHeights: TreeSet[Height]
 
 }
