+++ modules/explorer-core/src/main/scala/org/ergoplatform/uexplorer/plugin/Plugin.scala
@@ -3,8 +3,7 @@ package org.ergoplatform.uexplorer.plugin
 import org.apache.tinkerpop.gremlin.process.traversal.dsl.graph.GraphTraversalSource
 import org.ergoplatform.uexplorer.db.{BestBlockInserted, Block}
 import org.ergoplatform.uexplorer.node.ApiTransaction
-import org.ergoplatform.uexplorer.utxo.UtxoState
-import org.ergoplatform.uexplorer.{Address, BoxId, SortedTopAddressMap, TopAddressMap, TxId, Value}
+import org.ergoplatform.uexplorer.{Address, BoxId, SortedTopAddressMap, Storage, TopAddressMap, TxId, Value}
 
 import scala.collection.immutable.ListMap
 import scala.concurrent.Future
@@ -20,13 +19,13 @@ trait Plugin {
 
   def processMempoolTx(
     newTx: ApiTransaction,
-    utxoState: UtxoState,
+    storage: Storage,
     graphTraversalSource: Option[GraphTraversalSource]
   ): Future[Unit]
 
   def processNewBlock(
     newBlock: BestBlockInserted,
-    utxoState: UtxoState,
+    storage: Storage,
     graphTraversalSource: Option[GraphTraversalSource]
   ): Future[Unit]
 }
