+++ modules/mvstore/src/main/scala/org/ergoplatform/uexplorer/mvstore/kryo/AddressStatsCodec.scala
@@ -1,7 +1,4 @@
-package org.ergoplatform.uexplorer.utxo
-
-import org.ergoplatform.uexplorer.BlockMetadata
-import org.ergoplatform.uexplorer.db.DbCodec
+package org.ergoplatform.uexplorer.mvstore.kryo
 
 import com.esotericsoftware.kryo.Kryo
 import com.esotericsoftware.kryo.io.{ByteBufferOutput, Input}
@@ -10,25 +7,26 @@ import com.esotericsoftware.kryo.serializers.ImmutableCollectionsSerializers.Jdk
 import com.esotericsoftware.kryo.serializers.{ImmutableCollectionsSerializers, MapSerializer}
 import com.esotericsoftware.kryo.util.Pool
 import org.ergoplatform.uexplorer.db.BlockInfo
+import org.ergoplatform.uexplorer.mvstore.{DbCodec}
 import org.ergoplatform.uexplorer.{Address, BlockMetadata, Height}
 
 import java.nio.ByteBuffer
 import java.util
 import scala.util.Try
 
-object BlockMetadataCodec extends DbCodec[BlockMetadata] {
-  override def read(bytes: Array[Byte]): BlockMetadata = {
+object AddressStatsCodec extends DbCodec[Address.Stats] {
+  override def read(bytes: Array[Byte]): Address.Stats = {
     val input = new Input(bytes)
     val kryo  = KryoSerialization.pool.obtain()
-    try kryo.readObject(input, classOf[BlockMetadata])
+    try kryo.readObject(input, classOf[Address.Stats])
     finally {
       KryoSerialization.pool.free(kryo)
       input.close()
     }
   }
 
-  override def write(obj: BlockMetadata): Array[Byte] = {
-    val buffer = ByteBuffer.allocate(4096)
+  override def write(obj: Address.Stats): Array[Byte] = {
+    val buffer = ByteBuffer.allocate(256)
     val output = new ByteBufferOutput(buffer)
     val kryo   = KryoSerialization.pool.obtain()
     try kryo.writeObject(output, obj)
