+++ modules/mvstore/src/main/scala/org/ergoplatform/uexplorer/mvstore/kryo/BlockIdsCodec.scala
@@ -1,4 +1,4 @@
-package org.ergoplatform.uexplorer.utxo
+package org.ergoplatform.uexplorer.mvstore.kryo
 
 import com.esotericsoftware.kryo.Kryo
 import com.esotericsoftware.kryo.io.{ByteBufferOutput, Input}
@@ -6,29 +6,30 @@ import com.esotericsoftware.kryo.serializers.DefaultSerializers.CollectionsSingl
 import com.esotericsoftware.kryo.serializers.ImmutableCollectionsSerializers.JdkImmutableSetSerializer
 import com.esotericsoftware.kryo.serializers.{ImmutableCollectionsSerializers, MapSerializer}
 import com.esotericsoftware.kryo.util.Pool
-import org.ergoplatform.uexplorer.db.{BlockInfo, DbCodec}
-import org.ergoplatform.uexplorer.{Address, BlockMetadata, BoxId, Height, Value}
+import org.ergoplatform.uexplorer.db.BlockInfo
+import org.ergoplatform.uexplorer.mvstore.{DbCodec}
+import org.ergoplatform.uexplorer.{Address, BlockId, BlockMetadata, Height}
 
 import java.nio.ByteBuffer
 import java.util
 import scala.util.Try
 
-object ValueByBoxCodec extends DbCodec[java.util.Map[BoxId, Value]] {
-  override def read(bytes: Array[Byte]): java.util.Map[BoxId, Value] = {
+object BlockIdsCodec extends DbCodec[java.util.Set[BlockId]] {
+  override def read(bytes: Array[Byte]): java.util.Set[BlockId] = {
     val input = new Input(bytes)
     val kryo  = KryoSerialization.pool.obtain()
-    try kryo.readObject(input, classOf[util.HashMap[BoxId, Value]])
+    try kryo.readObject(input, classOf[util.HashSet[BlockId]])
     finally {
       KryoSerialization.pool.free(kryo)
       input.close()
     }
   }
 
-  override def write(valueByBoxId: java.util.Map[BoxId, Value]): Array[Byte] = {
-    val buffer = ByteBuffer.allocate(valueByBoxId.size() * 256)
+  override def write(blockIds: java.util.Set[BlockId]): Array[Byte] = {
+    val buffer = ByteBuffer.allocate(blockIds.size() * 128)
     val output = new ByteBufferOutput(buffer)
     val kryo   = KryoSerialization.pool.obtain()
-    try kryo.writeObject(output, valueByBoxId)
+    try kryo.writeObject(output, blockIds)
     finally {
       KryoSerialization.pool.free(kryo)
       output.close()
