+++ modules/mvstore/src/main/scala/org/ergoplatform/uexplorer/mvstore/kryo/BlockMetadataCodec.scala
@@ -1,4 +1,4 @@
-package org.ergoplatform.uexplorer.utxo
+package org.ergoplatform.uexplorer.mvstore.kryo
 
 import com.esotericsoftware.kryo.Kryo
 import com.esotericsoftware.kryo.io.{ByteBufferOutput, Input}
@@ -6,29 +6,30 @@ import com.esotericsoftware.kryo.serializers.DefaultSerializers.CollectionsSingl
 import com.esotericsoftware.kryo.serializers.ImmutableCollectionsSerializers.JdkImmutableSetSerializer
 import com.esotericsoftware.kryo.serializers.{ImmutableCollectionsSerializers, MapSerializer}
 import com.esotericsoftware.kryo.util.Pool
-import org.ergoplatform.uexplorer.db.{BlockInfo, DbCodec}
-import org.ergoplatform.uexplorer.{Address, BlockId, BlockMetadata, Height}
+import org.ergoplatform.uexplorer.db.BlockInfo
+import org.ergoplatform.uexplorer.mvstore.{DbCodec}
+import org.ergoplatform.uexplorer.{Address, BlockMetadata, Height}
 
 import java.nio.ByteBuffer
 import java.util
 import scala.util.Try
 
-object BlockIdsCodec extends DbCodec[java.util.Set[BlockId]] {
-  override def read(bytes: Array[Byte]): java.util.Set[BlockId] = {
+object BlockMetadataCodec extends DbCodec[BlockMetadata] {
+  override def read(bytes: Array[Byte]): BlockMetadata = {
     val input = new Input(bytes)
     val kryo  = KryoSerialization.pool.obtain()
-    try kryo.readObject(input, classOf[util.HashSet[BlockId]])
+    try kryo.readObject(input, classOf[BlockMetadata])
     finally {
       KryoSerialization.pool.free(kryo)
       input.close()
     }
   }
 
-  override def write(blockIds: java.util.Set[BlockId]): Array[Byte] = {
-    val buffer = ByteBuffer.allocate(blockIds.size() * 128)
+  override def write(obj: BlockMetadata): Array[Byte] = {
+    val buffer = ByteBuffer.allocate(4096)
     val output = new ByteBufferOutput(buffer)
     val kryo   = KryoSerialization.pool.obtain()
-    try kryo.writeObject(output, blockIds)
+    try kryo.writeObject(output, obj)
     finally {
       KryoSerialization.pool.free(kryo)
       output.close()
