+++ modules/mvstore/src/main/scala/org/ergoplatform/uexplorer/mvstore/kryo/ValueByBoxCodec.scala
@@ -1,4 +1,4 @@
-package org.ergoplatform.uexplorer.utxo
+package org.ergoplatform.uexplorer.mvstore.kryo
 
 import com.esotericsoftware.kryo.Kryo
 import com.esotericsoftware.kryo.io.{ByteBufferOutput, Input}
@@ -6,29 +6,30 @@ import com.esotericsoftware.kryo.serializers.DefaultSerializers.CollectionsSingl
 import com.esotericsoftware.kryo.serializers.ImmutableCollectionsSerializers.JdkImmutableSetSerializer
 import com.esotericsoftware.kryo.serializers.{ImmutableCollectionsSerializers, MapSerializer}
 import com.esotericsoftware.kryo.util.Pool
-import org.ergoplatform.uexplorer.db.{BlockInfo, DbCodec}
-import org.ergoplatform.uexplorer.{Address, BlockMetadata, Height}
+import org.ergoplatform.uexplorer.db.BlockInfo
+import org.ergoplatform.uexplorer.mvstore.{DbCodec}
+import org.ergoplatform.uexplorer.{Address, BlockMetadata, BoxId, Height, Value}
 
 import java.nio.ByteBuffer
 import java.util
 import scala.util.Try
 
-object AddressStatsCodec extends DbCodec[Address.Stats] {
-  override def read(bytes: Array[Byte]): Address.Stats = {
+object ValueByBoxCodec extends DbCodec[java.util.Map[BoxId, Value]] {
+  override def read(bytes: Array[Byte]): java.util.Map[BoxId, Value] = {
     val input = new Input(bytes)
     val kryo  = KryoSerialization.pool.obtain()
-    try kryo.readObject(input, classOf[Address.Stats])
+    try kryo.readObject(input, classOf[util.HashMap[BoxId, Value]])
     finally {
       KryoSerialization.pool.free(kryo)
       input.close()
     }
   }
 
-  override def write(obj: Address.Stats): Array[Byte] = {
-    val buffer = ByteBuffer.allocate(256)
+  override def write(valueByBoxId: java.util.Map[BoxId, Value]): Array[Byte] = {
+    val buffer = ByteBuffer.allocate(valueByBoxId.size() * 256)
     val output = new ByteBufferOutput(buffer)
     val kryo   = KryoSerialization.pool.obtain()
-    try kryo.writeObject(output, obj)
+    try kryo.writeObject(output, valueByBoxId)
     finally {
       KryoSerialization.pool.free(kryo)
       output.close()
